Fix the usage of cookie API
This fixes an infinite redirect on moodle when a cookie expires for some reason (for example due to moodle server failure).

Before that we were inadverently passing { domain: ... } to browser.cookies.set. This created a non-host-only cookie that got treated a separate cookie from the one we used .get on. They got sent to the server in UNSPECIFIED order and in the end it moodle couldn't remove the stale cookie that was extended by us. Now we acknowledge that the .get and .set have different semantics wrt to what object they receive and the double-cookie issue is gone.

It took several months to debug that...